- changed class name - added base contact menu
Changed the class name and include so that it makes sense. It's the
class of contact and it registers the 2 cpts of user and company.

Put the CPT's under the contacts menu so that they are somewhere during
development. Long term they will be hidden by default and you'll be able
to invoke them with a 'dev' constant or something fancy like that.