- basic saving
We are not saving all the fields yet but it does work to create a new
post type entry and the title looks right.

I need to work through the rest of the save functions and then hook up
the table so that it does pull the data from our users.